Move tests to test folder
