+++ ispyb-ejb/src/test/java/ispyb/server/common/test/services/ComparatorTest.java
@@ -17,7 +17,7 @@
  * Contributors : S. Delageniere, R. Leal, L. Launer, K. Levik, S. Veyrier, P. Brenchereau, M. Bodin, A. De Maria Antolinos
  ******************************************************************************************************************************/
 
-package ispyb.server.common.test;
+package ispyb.server.common.test.services;
 
 
 import ispyb.server.biosaxs.services.core.ExperimentScope;
